platform/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pybind11>=2.11.0
